Distro inheritance semantics
*** This is an experimental patch, not well tested and incomplete ***

The distro is always inherited from the base-level profile, even if
the 'distro' attribute is overridden at higher levels in the
inheritance tree.

This differs from other missing attributes, which are inherited from
the closest ancestor where the attribute is set.

This can be counter-intuitive when higher-level profile 'distro'
attributes are set with the expectation that the base-level distro
will be overridden.

This incomplete patch does part of the work to implement these changed
semantics.

Things to note:

Some of the changed code is hacking around Profile.get_parent(), which
returns the distro object if there's no self.parent.  I don't know the
code well enough to remove that behavior.

Tree objects have a new 'get_inherited_distro()' method that returns
the distro object set in the nearest ancestor (starting with itself).
This function should replace 'get_conceptual_parent()' when the
'conceptual parent' is expected to be a distro object.

Then utils.grab_tree() builds the tree with the top inherited distro
at the base instead of the base profile's distro.  When
utils.blender() runs __consolidate(), it picks up attributes from this
distro object.  (__consolidate() also no longer treats the
Profile.distro attribute specially, so it is inherited as other
attributes and matches the inherited distro object's name in 'report'
and 'dumpvars'.)

A known, major missing piece is many other places where 'distro =
profile.get_conceptual_parent()' should be replaced with
get_inherited_distro().