koan: Add storage pool refresh function
When an RBD volume is created with qemu-img, libvirt is not
automatically notified, and trying to attach the volume causes an
exception.  The storage_pool_refresh() function scans uses virsh to
scan pools, and any pool that contains one of the created disks is
refreshed so that libvirt knows the volume exists.